+++ synapseclient/__main__.py
@@ -1,68 +1,8 @@
 """
-***************************
-Synapse command line client
-***************************
+The Synapse command line client.
 
-The Synapse Python Client can be used from the command line via the **synapse** command.
-
-Installation
-============
-
-The command line client is installed along with `installation of the Synapse Python client \
-<http://python-docs.synapse.org/build/html/index.html#installation>`_.
-
-Help
-====
-
-For help, type::
-
-    synapse -h.
-
-For help on specific commands, type::
-
-    synapse [command] -h
-
-
-Optional arguments
-==================
-
-.. code-block:: shell
-
-    -h, --help            show this help message and exit
-    --version             show program's version number and exit
-    -u SYNAPSEUSER, --username SYNAPSEUSER
-                        Username used to connect to Synapse
-    -p SYNAPSEPASSWORD, --password SYNAPSEPASSWORD
-                        Password used to connect to Synapse
-
-Commands
-========
-  * **get**              - download an entity and associated data
-  * **sync**             - Synchronize files described in a manifest to Synapse
-  * **store**            - uploads and adds a file to Synapse
-  * **store-table**      - uploads a table to Syanpse given a csv
-  * **add**              - add or modify content to Synapse
-  * **mv**               - move a dataset in Synapse
-  * **cp**               - copy an entity/dataset in Synapse
-  * **associate**        - Associate local files with the files stored in Synapse so
-                           that calls to 'syntapse get' and 'synapse show' don't
-                           re-download the files, but use the already existing file.
-  * **delete**           - removes a dataset from Synapse
-  * **query**            - performs SQL like queries on Synapse
-  * **submit**           - submit an entity for evaluation
-  * **show**             - displays information about a Entity
-  * **cat**              - prints a dataset from Synapse
-  * **list**             - List Synapse entities contained by the given Project or
-                           Folder. Note: May not be supported in future versions of
-                           the client.
-  * **set-provenance**   - create provenance records
-  * **get-provenance**   - show provenance records
-  * **set-annotations**  - create annotations
-  * **get-annotations**  - show annotations
-  * **create**           - Creates folders or projects on Synapse
-  * **onweb**            - opens Synapse website for Entity
-  * **login**            - login to Synapse and (optionally) cache credentials
-  * **test-encoding**    - test character encoding to help diagnose problems
+For a description of its usage and parameters, see its documentation:
+https://python-docs.synapse.org/build/html/CommandLineClient.html
 """
 import argparse
 import collections.abc
@@ -545,9 +485,9 @@ def build_parser():
     parser_get.add_argument('--downloadLocation', metavar='path', type=str, default="./",
                             help='Directory to download file to [default: %(default)s].')
     parser_get.add_argument('--multiThreaded', action='store_true',
-                            default=False, help='Download file using a multiple threaded implementation. '
-                                                'This flag will be removed in the future when multi-threaded download '
-                                                'is deemed fully stable and becomes the default implementation.')
+                            default=True, help='Download file using a multiple threaded implementation. '
+                            'This flag will be removed in the future when multi-threaded download '
+                            'is deemed fully stable and becomes the default implementation.')
     parser_get.add_argument('id', metavar='syn123', nargs='?', type=str,
                             help='Synapse ID of form syn123 of desired data object.')
     parser_get.set_defaults(func=get)